cof editor: moved code responsible for editing into hswidget/cofeditor.go; created PopUpDialog widget
